+++ libraries/SXMac-Bridging-Header.h
@@ -5,3 +5,4 @@
 #import "pathsManager.h" //manage paths for SXMac 
 #import "alertView.h" //alert in Objective-C
 #import "SXMacMethods.h"
+#import "sortByFileSize.h"