+++ libraries/sortByFileSize.h
@@ -7,9 +7,14 @@
 
 #import <Foundation/Foundation.h>
 
+#import <sys/stat.h>
+
 NS_ASSUME_NONNULL_BEGIN
 
 @interface sortByFileSize : NSObject
+
+- (NSArray<NSString *> *)sortByFileSize:(NSArray<NSString *> *)files;
+
 
 @end